requirements.txt
Django==1.11.16
django-filter==2.0.0
djangorestframework==3.9.0
Markdown==3.0.1
Pillow==5.3.0
psycopg2==2.7.5
psycopg2-binary==2.7.7
pytz==2018.7
